+++ src/device/tsp/Tsp/SessionBaseExtensions/SessionBaseMethods_node.cs
@@ -4,6 +4,22 @@ namespace cc.isr.VI.Tsp.SessionBaseExtensions;
 
 public static partial class SessionBaseMethods
 {
+
+    /// <summary>
+    /// A Pith.SessionBase extension method that queries controller node serial number.
+    /// </summary>
+    /// <remarks>   2025-05-26. </remarks>
+    /// <param name="session">  The session to act on. </param>
+    /// <returns>   The controller node serial number. </returns>
+    public static string QueryControllerNodeSerialNumber( this Pith.SessionBase session )
+    {
+        if ( session is null ) throw new ArgumentNullException( nameof( session ) );
+        if ( !session.IsDeviceOpen ) throw new InvalidOperationException( $"{nameof( session )} is not open." );
+
+        session.TraceLastAction( "fetching controller node serial number." );
+        return session.QueryTrimEnd( cc.isr.VI.Syntax.Tsp.LocalNode.SerialNumberFormattedQueryCommand );
+    }
+
     /// <summary>   A <see cref="Pith.SessionBase"/>  extension method that queries controller node number. </summary>
     /// <remarks>   2025-04-26. </remarks>
     /// <exception cref="ArgumentNullException">    Thrown when one or more required arguments are