+++ changelog.md
@@ -2,7 +2,16 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9275]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9279]: https://www.github.com/atecoder/dn.vi.ivi
+
+## [8.1.9279] - 2025-05-28
+- VI.Device Access Subsystem Base:
+  - Consolidate the registration code.
+  - Add IsCertified.
+- Device Tsp Session Base Extensions:
+  - Add method to query the controller node serial number.
+- Script Info Collection:
+  - Add methods to define the deploy file name using a model and major version.
 
 ## [8.1.9275] - 2025-05-24
 - Device Tsp Script